Passing tests
written a function that passes the test that I have written for this. Have had to change it as I have tried to do it in ES6 but at the current moment have not been able to do it properly. Would like to find a way to do this, but at the current moment will write this in ES5 just to get everything working.